refactor(018/US2): split npm.rs into npm/ submodule
Split mikebom-cli/src/scan_fs/package_db/npm.rs (1616 LOC) into
five files under npm/:

  npm/mod.rs           ~525 LOC orchestrator + read() + NpmError +
                                project-root walker + NpmIntegrity
                                + base64 + build_npm_purl
  npm/package_lock.rs  ~280 LOC package-lock.json v2/v3 parser +
                                derive_name_from_path_key
  npm/pnpm_lock.rs     ~190 LOC pnpm-lock.yaml parser + parse_pnpm_key
  npm/walk.rs          ~330 LOC node_modules flat walker +
                                read_root_package_json +
                                classify_npm_source
  npm/enrich.rs        ~120 LOC author backfill from installed
                                package.json files

Visibility ladder:
  - pub fn read, pub enum NpmError re-exported from npm/mod.rs
    unchanged
  - pub(super) fn for cross-sibling parser entry points
    (read_package_lock, read_pnpm_lock, read_node_modules,
    read_root_package_json, enrich_entries_with_installed_authors,
    extract_author_string)
  - pub(super) NpmIntegrity reachable to package_lock + pnpm_lock
    siblings
  - private fn for parser-internal helpers (parse_pnpm_key,
    derive_name_from_path_key, person_from_value, classify_npm_source)
  - tests inside #[cfg(test)] mod tests at the same level as the
    code they exercise (per FR-007). 39 tests distributed across:
    enrich (6), package_lock (8), pnpm_lock (3), walk (5), mod (17)

Verified:
  - cargo +stable clippy --workspace --all-targets -- -D warnings
    passes clean
  - 39 npm tests + dependents all pass
  - All 27 byte-identity goldens unchanged: regen produces zero
    `git diff`

US3 (binary split) remains.
